Removed all truffle related files - Truffle configs - Migration files - testing scripts
Migrate coverage tool from truffle to hardhat

Add hardhat test script

Update README.md to use node 14 and yarn

Fix broken unit tests by switching to from npm to yarn
