+++ test/gas.js
@@ -3,7 +3,7 @@ require('should');
 const BigNumber = require('bignumber.js');
 
 const helpers = require('./helpers');
-const { privateKeyForAccount } = require('../testrpc/accounts');
+const { privateKeyForAccount } = require('./helpers');
 
 // Used to build the solidity tightly packed buffer to sha3, ecsign
 const util = require('ethereumjs-util');
